django_multi_schema/django_multiple_schema/multiSchema/requirements.txt
apispec==6.0.2
asgiref==3.6.0
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
dataclasses-jsonschema==2.16.0
Django==3.2
django-dataclass-autoserialize==1.0.1
django-dataclasses==0.2.2
django-extensions==3.2.1
django-ipware==4.0.2
django-rest-framework==0.1.0
django-structlog==4.1.1
django-tenant-schemas==1.11.0
djangorestframework==3.14.0
djangorestframework-dataclasses==1.2.0
drf-spectacular==0.25.1
drf-yasg==1.21.5
idna==3.4
importlib-resources==5.10.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mypy-extensions==1.0.0
numpy==1.24.2
ordered-set==4.1.0
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
pkgutil-resolve-name==1.3.10
platformdirs==3.0.0
psycopg2-binary==2.9.5
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
structlog==22.3.0
tomli==2.0.1
typing-extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
zipp==3.13.0
